Added two classes for spatial indexing of a "pcd" numpy array; "Grid" and "QuadTree".
For a small number of grid-cells "Grid" is a clear winner, however the time taken to create the "Grid" is dominated by the spatial resolution/number of cells in the grid. It's possible the author has just done something stupid.

For more general usage the QuadTree approach has better performance. Subsequent commit will remove the Grid class altogether.